修正
book和user的404 删除书检查是否为作者

+++ src/servlets/Book/DeleteBookServlet.java
@@ -1,6 +1,7 @@
 package servlets.Book;
 
 import Util.FileUtil;
+import model.Book;
 import service.BookService;
 import service.impl.BookServiceImpl;
 
@@ -22,26 +23,33 @@ public class DeleteBookServlet extends HttpServlet {
 
     public void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        BookService bookService = new BookServiceImpl();
-        // 数据库删除
-        int bookID = Integer.parseInt(request.getParameter("bookID"));
-        bookService.delete(bookID, (String) request.getSession().getAttribute("username"));
-        // 删除文件
         try {
+            BookService bookService = new BookServiceImpl();
+            // 数据库删除
+            int bookID = Integer.parseInt(request.getParameter("bookID"));
+            Book book = bookService.find(bookID);
+            if (!request.getSession().getAttribute("username").equals(book.getChiefEditor())) {
+                throw new Exception("用户不是作者");
+            }
+            bookService.delete(bookID, (String) request.getSession().getAttribute("username"));
+            // 删除文件
             File cover = new File(this.getServletContext().getRealPath("/resources/cover/" + bookID + ".jpg")); // cover 是 jpg 文件
-            File book = new File(this.getServletContext().getRealPath("/resources/book/" + bookID)); // book 是目录
+            File bookFolder = new File(this.getServletContext().getRealPath("/resources/book/" + bookID)); // book 是目录
             if (cover.exists() && cover.isFile())
                 cover.delete();
-            if (book.exists() && book.isDirectory()) // TODO 当前无法删除该文件夹
+            if (bookFolder.exists() && bookFolder.isDirectory()) // TODO 当前无法删除该文件夹
             {
-                if (FileUtil.deleteDir(book))
-                    System.out.println("DeleteBookServlet: 删除目录成功 " + book.getPath());
+                if (FileUtil.deleteDir(bookFolder))
+                    System.out.println("DeleteBookServlet: 删除目录成功 " + bookFolder.getPath());
             }
+            // 删除本书后，重定向回首页
+            response.sendRedirect("/home");
+        } catch (NullPointerException e) {
+            response.sendError(404);
         } catch (Exception e) {
             System.out.println("DeleteBookServlet: 删除书目文件失败");
             e.printStackTrace();
+            response.sendError(500);
         }
-        // 删除本书后，重定向回首页
-        response.sendRedirect("/home");
     }
 }
